refactor: remove exception emission when logzion has a connection error
