+++ lib/common/logging/logging.dart
@@ -2,5 +2,4 @@ export 'callkeep_logs.dart';
 export 'filtered_logz_io_appender.dart';
 export 'logzio_logging_service.dart';
 export 'remote_formatter.dart';
-export 'remote_log_filter.dart';
 export 'remote_logging_service.dart';
